Split handleTransaction and handleHeaders into separate files and fix tests compilation.
